Abstract actions

+++ frontend/src/src/app/section/auth/token-storage.service.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
 import {AuthoritiesResponse} from "./authorities-response";
+import * as jwt_decode from 'jwt-decode';
 
 const TOKEN_KEY = 'AuthToken';
+const TOKEN_EXPIRED = 'TokenExpired'
 const EMAIL_KEY = 'AuthEmail';
 const AUTHORITIES_KEY = 'AuthAuthorities';
 const USERID_KEY = 'AuthUserId';
@@ -14,8 +16,10 @@ export class TokenStorageService {
   constructor() { }
 
   signOut() {
-    window.localStorage.clear();
-    window.location.replace('')
+    if (window.localStorage.length !== 0) {
+      window.localStorage.clear();
+      window.location.replace('');
+    }
   }
 
   public saveToken(accessToken: string) {
@@ -25,6 +29,28 @@ export class TokenStorageService {
 
   public getToken(): string {
     return localStorage.getItem(TOKEN_KEY);
+  }
+
+  private getTokenExpirationDate(token: String) {
+    const decoded = jwt_decode(token);
+
+    if(decoded.exp === undefined) {
+      return null;
+    }
+
+    const date = new Date(0);
+    date.setUTCSeconds(decoded.exp);
+    return date;
+  }
+
+  public isTokenExpired(token?: string): boolean {
+    if(!token) token = this.getToken();
+    if(!token) return true;
+
+    const date = this.getTokenExpirationDate(token);
+    console.log(date);
+    if(date === undefined) return false;
+    return !(date.valueOf() > new Date().valueOf());
   }
 
   public saveEmail(email: string) {
